Add deploy specs to script archives
This provides a placeholder for separate deployment metadata, as opposed to archive authoring
metadata. The application can create, or modify the deployment metadata as it sees fit, allowing the
loader or executor to treat the compiled module in various custom ways. Note that the application,
or runtime could modify deployment metadata for a published script archive, whereas the authoring
metadata is intended to be treated as immutable.