Modify how compiled classes are injected into the module class loader
The previous method did not work when a class being injected was a subclass of another class yet to
be injected. A variation of the problem is the case where the superclass is in a different package.

The new methodology is to ask the module's classloader to refresh its resource loader paths (which
will pick up the packages created in the compilation root directory), and then walk the compilation
tree to look for class files, and load them iteratively.

Alternative methods are possible, which will involve implementing a custom findClass() in
JBossModuleClassloader.